Fixed all swiftlint errors that were manageable.
